@@ -12,53 +12,52 @@
   },
   "private": true,
   "dependencies": {
-    "@agm/core": "1.0.0-beta.5",
-    "@angular/animations": "^7.2.15",
-    "@angular/cdk": "github:angular/cdk-builds",
-    "@angular/common": "^7.2.15",
-    "@angular/compiler": "^7.2.15",
-    "@angular/core": "^7.2.15",
-    "@angular/forms": "^7.2.15",
-    "@angular/http": "^7.2.15",
-    "@angular/platform-browser": "^7.2.15",
-    "@angular/platform-browser-dynamic": "^7.2.15",
-    "@angular/router": "^7.2.15",
-    "ag-grid-angular": "^20.2.0",
-    "ag-grid-community": "^20.2.0",
-    "core-js": "^3.0.0",
-    "express": "^4.16.4",
+    "@agm/core": "~1.0.0-beta.6",
+    "@angular/animations": "^8.1.0",
+    "@angular/cdk": "~8.0.2",
+    "@angular/common": "~8.1.0",
+    "@angular/compiler": "~8.1.0",
+    "@angular/core": "~8.1.0",
+    "@angular/forms": "~8.1.0",
+    "@angular/material": "^8.0.2",
+    "@angular/platform-browser": "~8.1.0",
+    "@angular/platform-browser-dynamic": "~8.1.0",
+    "@angular/router": "~8.1.0",
+    "@auth0/angular-jwt": "^2.1.1",
+    "ag-grid-angular": "^21.0.1",
+    "ag-grid-community": "^21.0.1",
+    "core-js": "^3.1.4",
+    "express": "^4.17.1",
     "font-awesome": "^4.7.0",
     "googlemaps": "^1.12.0",
     "moment": "^2.24.0",
     "ng2-bootstrap-modal": "^1.0.1",
-    "rxjs": "^6.3.3",
+    "rxjs": "^6.5.2",
     "rxjs-compat": "^6.5.2",
     "timespan": "^2.3.0",
     "typescript-dotnet-umd": "^4.11.2",
-    "zone.js": "^0.9.0"
+    "zone.js": "^0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.13.8",
-    "@angular/cli": "^7.3.9",
-    "@angular/compiler-cli": "^7.2.12",
-    "@angular/language-service": "^7.2.12",
-    "@babel/preset-env": "^7.4.3",
-    "@types/googlemaps": "^3.30.19",
-    "@types/jasmine": "~3.3.12",
-    "@types/jasminewd2": "~2.0.6",
-    "@types/node": "~11.13.0",
-    "codelyzer": "5.0.0",
-    "jasmine-core": "~3.4.0",
+    "@angular-devkit/build-angular": "^0.801.0",
+    "@angular/cli": "^8.1.0",
+    "@angular/compiler-cli": "^8.1.0",
+    "@angular/language-service": "^8.1.0",
+    "@types/googlemaps": "^3.36.7",
+    "@types/jasmine": "~2.8.8",
+    "@types/jasminewd2": "~2.0.3",
+    "@types/node": "^11.13.17",
+    "codelyzer": "5.1.0",
+    "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "^4.0.1",
+    "karma": "~4.0.0",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-cli": "~2.0.0",
-    "karma-coverage-istanbul-reporter": "^2.0.5",
-    "karma-jasmine": "~2.0.1",
-    "karma-jasmine-html-reporter": "^1.4.0",
-    "protractor": "~5.4.2",
-    "ts-node": "~8.0.3",
-    "tslint": "~5.15.0",
-    "typescript": "^3.2.4"
+    "karma-coverage-istanbul-reporter": "~2.0.1",
+    "karma-jasmine": "~1.1.2",
+    "karma-jasmine-html-reporter": "^0.2.2",
+    "protractor": "~5.4.0",
+    "ts-node": "~7.0.0",
+    "tslint": "~5.11.0",
+    "typescript": "^3.4.5"
   }
 }
