@@ -1,11 +1,8 @@
-import { Component, OnInit, ViewEncapsulation , Input,Output,EventEmitter} from '@angular/core';
-
-import { Subject } from 'rxjs/Subject';
-import { Observable } from 'rxjs/Observable';
+import { Component, AfterViewInit, ViewEncapsulation , Input,Output,EventEmitter} from '@angular/core';
+import { Observable } from 'rxjs';
 import { ProfesionalService, CommonService } from '../../../service/index';
 import { ProfesionalBE, PersonBE,  GetProfesionalRes, ResourceSchedulingBE,User,IContextInformation, IParam, Param, CommonValuesEnum, TipoParametroEnum, CommonParams, HealtConstants ,contextInfo} from '../../../model/index';
 import { FormGroup } from '@angular/forms';
-import { ViewChild, ElementRef, Renderer2, AfterContentInit, AfterViewInit } from '@angular/core';
 import { ActivatedRoute } from "@angular/router";
 import { ServiceError } from '../../../model/common.model';
 import { ControlContainer, NgForm } from '@angular/forms';
