@@ -4,9 +4,9 @@ import { ProfesionalBE, Profesional_FullViewBE, ProfesionalesGridBE,IContextInfo
 import {TipoParametroEnum} from '../../../model/common.constants'
 
 //permmite cambiar la variable obsevada
-import { Subject } from 'rxjs/Subject';
+
 //permite observar
-import { Observable } from 'rxjs/Observable';
+import { Observable } from 'rxjs';
 
 
 // rich grid and rich grid declarative
