@@ -3,7 +3,7 @@ import { ControlContainer, NgForm } from '@angular/forms';
 
 import { User, ServiceError, Rol } from "../../model/common.model";
 import { ProfesionalService }  from '../../service/index';
-import { Observable } from "rxjs/Observable";
+import { Observable } from "rxjs";
 
 class UserSession
 {
